class/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
black==24.4.2
boto3==1.35.45
botocore==1.35.45
CacheControl==0.14.3
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clicksend-client==5.0.78
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
decorator==5.1.1
distlib==0.3.8
Django==5.0.7
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-filter==24.2
django-multiselectfield==0.1.13
django-redis==5.4.0
django-storages==1.14.4
django-stubs-ext==5.0.2
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-standardized-errors==0.14.0
drf-yasg==1.21.7
et_xmlfile==2.0.0
filelock==3.15.4
firebase-admin==6.6.0
flake8==7.1.0
flower==2.0.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
humanize==4.10.0
hyperlink==21.0.0
identify==2.6.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
isort==5.13.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
lxml==5.3.0
mailersend==0.5.8
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.7.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==4.3.1
pypng==0.20220715.0
python-bidi==0.6.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==8.2
ratelim==0.1.6
redis==5.0.7
referencing==0.35.1
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.3
sentry-sdk==2.14.0
service-identity==24.1.0
setuptools==80.7.1
six==1.16.0
sqlparse==0.5.1
stripe==10.6.0
svglib==1.5.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
twilio==9.3.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
xero_python==9.0.0
xhtml2pdf==0.2.16
yarl==1.13.0
zope.interface==7.0.3
